Separate metrics test configuration from tests
