+++ Tests/ApodiniObserveTests/ApodiniMetricsTests.swift
@@ -1,20 +1,18 @@
-//                   
+//
 // This source file is part of the Apodini open source project
 //
 // SPDX-FileCopyrightText: 2019-2021 Paul Schmiedmayer and the Apodini project authors (see CONTRIBUTORS.md) <paul.schmiedmayer@tum.de>
 //
 // SPDX-License-Identifier: MIT
-//              
+//
 
-import XCTest
-import XCTApodini
-import XCTApodiniNetworking
-@testable import CoreMetrics
-@testable import Metrics
-@testable import ApodiniObserve
-import ApodiniHTTP
 @testable import Apodini
-@testable import MetricsTestUtils
+import ApodiniHTTP
+import ApodiniObserve
+import MetricsTestUtils
+import XCTApodini
+import XCTApodiniObserve
+import XCTest
 
 // swiftlint:disable closure_body_length
 public class ApodiniMetricsTests: XCTestCase {
@@ -58,7 +56,7 @@ public class ApodiniMetricsTests: XCTestCase {
                                         )
         )
         
-        app = Self.configureMetrics(app, metricsConfiguration: config)
+        app = XCTApodiniObserve.configureMetrics(app, metricsConfiguration: config)
         
         let visitor = SyntaxTreeVisitor(modelBuilder: SemanticModelBuilder(app))
         content.accept(visitor)
@@ -71,45 +69,6 @@ public class ApodiniMetricsTests: XCTestCase {
         app.shutdown()
         
         XCTAssertApodiniApplicationNotRunning()
-    }
-    
-    // Copied from the source code of ApodiniObserve to bootstrap the MetricsSystem internally
-    // (required for the tests, as the MetricsSystem only allows to be configured once per process)
-    public static func configureMetrics(_ app: Apodini.Application, metricsConfiguration: MetricsConfiguration) -> Apodini.Application {
-        // Bootstrap all passed MetricHandlers
-        MetricsSystem.bootstrapInternal(
-            MultiplexMetricsHandler(
-                factories: metricsConfiguration.metricHandlerConfigurations.map { $0.factory }
-            )
-        )
-        
-        if !app.checkRegisteredExporter(exporterType: ObserveMetadataExporter.self) {
-            // Instanciate exporter
-            let metadataExporter = ObserveMetadataExporter(app, metricsConfiguration)
-            
-            // Insert exporter into `InterfaceExporterStorage`
-            app.registerExporter(exporter: metadataExporter)
-        }
-        
-        metricsConfiguration.metricHandlerConfigurations.forEach { metricHandlerConfiguration in
-            if let metricPullHandlerConfiguration = metricHandlerConfiguration as? MetricPullHandlerConfiguration {
-                let endpoint = metricPullHandlerConfiguration.endpoint.hasPrefix("/")
-                                ? metricPullHandlerConfiguration.endpoint
-                                : "/\(metricPullHandlerConfiguration.endpoint)"
-                app.httpServer.registerRoute(.GET, endpoint.httpPathComponents) { req -> EventLoopFuture<String> in
-                    metricPullHandlerConfiguration.collect(req.eventLoop.makePromise(of: String.self))
-                }
-                
-                // Inform developer about which MetricsHandler serves the metrics data on what endpoint
-                app.logger.info("Metrics data of \(metricPullHandlerConfiguration.factory.self) served on \(metricPullHandlerConfiguration.endpoint)")
-            }
-        }
-        
-        // Write configuration to the storage
-        app.storage.set(MetricsConfiguration.MetricsStorageKey.self,
-                        to: MetricsConfiguration.MetricsStorageValue(configuration: metricsConfiguration))
-        
-        return app
     }
     
     struct Greeter: Handler {
